# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Companies",
        "version": "1.0.0"
    },
    "paths": {
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get Companies",
                "operationId": "54f966c056050f9db4fff6ee5a45a76a",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "parent_company_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "example name"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create Company",
                "operationId": "00e2db7b0f71c32b377f4dfc4dd685a8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "parent_company_id": {
                                                "type": "number"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "parent_company_id": "1",
                                    "name": "example name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "parent_company_id": {
                                            "type": "number",
                                            "example": "title"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get Company",
                "operationId": "01c261c039e4b27beeb888ef46bd86db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "parent_company_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update Company",
                "operationId": "6a1d7a4a2b9cbdfe7749cef78e6cec57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "parent_company_id": {
                                                "type": "number"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "parent_company_id": "1",
                                    "name": "example name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "parent_company_id": {
                                            "type": "number",
                                            "example": "title"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete Company",
                "operationId": "49d767f608bffc95ad39368bf1766693",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete company success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}